requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
pillow
unidecode
datasets
peft
